refactor(421): extract md + json host families; finish heap.py with shared collection helpers
Step 5 of the codegen/api.py decomposition (#421).  The Markdown (#9.7.3) and JSON families move out of execute() into vera/runtime/md.py (register_md(linker)) and vera/runtime/json.py (register_json(linker, ops_used) -- it has per-op guards).  Extracting json surfaced a gap the heap layer left behind: a block of pure marshalling helpers (_write_i64, _write_f64, _read_i32, _read_f64, _alloc_option_some_{i64,i32,f64}, _alloc_array_of_{i64,i32,f64}) was defined under a conditional deep in execute() and shared by all five collection families.  These are caller-parameterised, so they relocate unchanged to vera/runtime/heap.py; the still-inline Map/Set/Decimal/HTML blocks and the new json.py import them.  _VAL_WASM_TYPES and the _host_store_refs-bound host_decref_handle/host_attach_bucket stay inline (they go with the collections layer).  api.py 2,884->2,592 lines.  mypy + ruff clean; 1,233 codegen/characterization + 63 md/html tests green.
